% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps_get.R
\name{get_bed_overlaps}
\alias{get_bed_overlaps}
\title{Get overlaps of data.frame rows and BED files.}
\usage{
get_bed_overlaps(df, dir = ".", col_itemRgb = 0, verbose = F)
}
\arguments{
\item{df}{Data.frame.
Data.frame with at least these columns: CHROM, POS, ID.}

\item{dir}{Character.
String specifying the directory with BED files. This method assumes
there is not header to the BED files that must be removed.}

\item{col_itemRgb}{Numeric.
Specifies the column of itemRgb values within a BED file.
In a strict BED file, this would be column 9
    (see \url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}).
If \code{col_itemRgb == 0}, RGB values will not be appended to the
output, as described below in the return documentation of this function.}

\item{verbose}{Logical.
If true, print extra output (e.g., on bed files being read).}
}
\value{
Data.frame.
    The input data.frame with additional bed* columns named for each BED
    file in the input directory (e.g., bed.<file_name1>). The value of each
    column is the 'name' entry of the BED file (column 4) corresponding to
    the genomic segment that overlaps the input data.
    If \code{col_itemRgb != 0}, then the output values will have the RGB
    values from the BED file appended to them using a '::' character (e.g,
    <bed_file.name_value> \code{->}
    <bed_file.name_value>::<bed_file.RGB_value>).
}
\description{
\code{get_bed_overlaps} finds overlaps between the rows of a data.frame and
every BED file within a directory. If one row does not overlap anything in
the BED files, it is still returned.
}
