% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlaps_plot.R
\name{plot_overlaps}
\alias{plot_overlaps}
\title{Plot overlaps.}
\usage{
plot_overlaps(df, xid_solid_line = NULL, varshney_chrhmm = FALSE)
}
\arguments{
\item{df}{Data.frame.
Data.frame with at least these columns: CHROM, POS, ID, 'bed.*'.
The 'bed.*' column names define y-axis labels,
'bed.*' column values define the plot colors,
and the ID column values define the x-axis labels.}

\item{xid_solid_line}{Character vector.
A solid verticle line will be added to the plot at any x-axis value
that is contained within this character vector.}

\item{varshney_chrhmm}{Logical.
If \code{TRUE}, the function assumes the values of the 'bed.*' columns
are chromatin states from Varsheny et al 2017
(\url{https://doi.org/10.1073/pnas.1621192114}) and will format the
labels to make a more polished plot.}
}
\value{
List.
    A list with the following elements:
    \describe{
        \item{plt}{object from ggplot2}
        \item{df}{data.frame of the plot data}
    }
}
\description{
\code{plot_overlaps} plots the data.frame returned by
\code{\link{get_bed_overlaps}}.
}
