% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rgb2hex}
\alias{rgb2hex}
\title{RGB to HEX.}
\usage{
rgb2hex(x)
}
\arguments{
\item{x}{Character.
RGB string separated by commas.}
}
\value{
Character.
    A HEX value string.
}
\description{
\code{rgb2hex} converts RGB values to HEX values.
}
\examples{
rgb2hex('255,195,77')

}
