% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variantBedOverlap.R
\docType{package}
\name{variantBedOverlap}
\alias{variantBedOverlap}
\alias{variantBedOverlap-package}
\title{variantBedOverlap: Calculates overlap of variants in LD with bed files.}
\description{
Gets variants in LD with tag variant. Calculates the overlap of all variants
with all bed files in a directory. Saves the overlaps data.frame. Plots the
overlaps.
}
\section{variantBedOverlap functions}{

    \code{rgb2hex}
    \code{get_bed_overlaps}
    \code{plot_overlaps}
}

